Preview in its own window, other minor changes
