+++ src/components/infopopups/AppellantInitialPopup.js
@@ -5,7 +5,7 @@ import {
     SERVE_EACH_RESPONDENT_MSG,
     FILE_STEP_TWO_MSG
 } from "../../helpers/constants";
-import InfoPopupWrapper from "./common/InfoPopupWrapper";
+import InfoPopup from "./common/InfoPopup";
 
 class AppellantInitialPopup extends Component {
 
@@ -92,7 +92,7 @@ class AppellantInitialPopup extends Component {
     render() {
 
         return (
-            <InfoPopupWrapper
+            <InfoPopup
                 title="Initial Documents"
                 helpType="appellant"
                 close={this.props.close}
@@ -102,7 +102,7 @@ class AppellantInitialPopup extends Component {
                         Would you like to start your appeal?</div>
                 </div>
                 {this.props.getSections(this.sections)}
-            </InfoPopupWrapper>
+            </InfoPopup>
         );
     }
