+++ src/forms/Form2Preview.css
@@ -10,3 +10,13 @@
 #form2-preview td {
     padding: 5px;
 }
+.preview-form {
+    padding-right: 15px;
+}
+.preview-title {
+    background-color: var(--header-background);
+    color: white;
+    margin: auto;
+    padding: 20px;
+    font-size: 2.5rem;
+}
