+++ tests/components/infopopups/common/info.popups.spec.js
@@ -3,7 +3,7 @@ import ApplyForLeavePopup from "../../../../src/components/infopopups/ApplyForLe
 require('../../../support/enzyme.setup');
 import React from 'react';
 import { mount } from 'enzyme';
-import InfoPopup from '../../../../src/components/infopopups/common/InfoPopup';
+import InfoPopupFactory from '../../../../src/components/infopopups/common/InfoPopupFactory';
 import AppellantFactumPopup from '../../../../src/components/infopopups/AppellantFactumPopup';
 import ReplyBookPopup from '../../../../src/components/infopopups/ReplyBookPopup';
 import HearingPopup from '../../../../src/components/infopopups/HearingPopup';
@@ -25,12 +25,12 @@ import CompleteAppealPopup from "../../../../src/components/infopopups/CompleteA
 import RespondToAppealHearingPopup from "../../../../src/components/infopopups/RespondToAppealHearingPopup";
 import RespondToLeaveToAppealHearingPopup from "../../../../src/components/infopopups/RespondToLeaveToAppealHearingPopup"
 
-describe('InfoPopup', ()=> {    
+describe('InfoPopupFactory', ()=> {
 
     describe('factory', ()=>{
 
         test('returns null when hidden', ()=>{
-            let popup = mount(<InfoPopup />);
+            let popup = mount(<InfoPopupFactory />);
             let view = popup.instance().render();
 
             expect(view).toEqual(null);
@@ -43,21 +43,21 @@ describe('InfoPopup', ()=> {
         let respondToLeaveRefused = ()=>{};
 
         test('can build AppellantFactumPopup', ()=>{            
-            let popup = mount(<InfoPopup show="true" type="factum" close={close}/>).instance();
+            let popup = mount(<InfoPopupFactory show="true" type="factum" close={close}/>).instance();
             let view = popup.render();
 
             expect(view).toEqual(<AppellantFactumPopup close={close} getSections={popup.getSections}/>);
         });
         
         test('can build ReplyBookPopup', ()=>{            
-            let popup = mount(<InfoPopup show="true" type="replybook" close={close}/>).instance();
+            let popup = mount(<InfoPopupFactory show="true" type="replybook" close={close}/>).instance();
             let view = popup.render();
 
             expect(view).toEqual(<ReplyBookPopup close={close} getSections={popup.getSections}/>);
         });
         
         test('can build HearingPopup for respondent', ()=>{
-            let popup = mount(<InfoPopup show="true" type="respondenthearing" close={close}/>).instance();
+            let popup = mount(<InfoPopupFactory show="true" type="respondenthearing" close={close}/>).instance();
             let view = popup.render();
 
             expect(view).toEqual(<HearingPopup
@@ -71,7 +71,7 @@ describe('InfoPopup', ()=> {
         });
 
         test('can build HearingPopup for appellant', ()=>{
-            let popup = mount(<InfoPopup show="true" type="appellanthearing" close={close}/>).instance();
+            let popup = mount(<InfoPopupFactory show="true" type="appellanthearing" close={close}/>).instance();
             let view = popup.render();
 
             expect(view).toEqual(
@@ -86,112 +86,112 @@ describe('InfoPopup', ()=> {
         });
         
         test('can build CourtOrderPopup', ()=>{            
-            let popup = mount(<InfoPopup show="true" type="courtorder" close={close}/>).instance();
+            let popup = mount(<InfoPopupFactory show="true" type="courtorder" close={close}/>).instance();
             let view = popup.render();
 
             expect(view).toEqual(<CourtOrderPopup close={close} getSections={popup.getSections}/>);
         });
         
         test('can build RespondentFactumPopup', ()=>{            
-            let popup = mount(<InfoPopup show="true" type="respondentfactum" close={close}/>).instance();
+            let popup = mount(<InfoPopupFactory show="true" type="respondentfactum" close={close}/>).instance();
             let view = popup.render();
 
             expect(view).toEqual(<RespondentFactumPopup close={close} getSections={popup.getSections}/>);
         });
         
         test('can build CrossAppealPopup', ()=>{            
-            let popup = mount(<InfoPopup show="true" type="crossappeal" close={close}/>).instance();
+            let popup = mount(<InfoPopupFactory show="true" type="crossappeal" close={close}/>).instance();
             let view = popup.render();
 
             expect(view).toEqual(<CrossAppealPopup close={close} getSections={popup.getSections}/>);
         });
         
         test('can build AppearancePopup', ()=>{            
-            let popup = mount(<InfoPopup show="true" type="leave-application-appearance" close={close}/>).instance();
+            let popup = mount(<InfoPopupFactory show="true" type="leave-application-appearance" close={close}/>).instance();
             let view = popup.render();
 
             expect(view).toEqual(<AppearancePopup close={close} getSections={popup.getSections} noticeType="Notice of Application for Leave to Appeal" />);
         });
         
         test('can build AppearancePopup with notice', ()=>{            
-            let popup = mount(<InfoPopup show="true" type="appearance" close={close}/>).instance();
+            let popup = mount(<InfoPopupFactory show="true" type="appearance" close={close}/>).instance();
             let view = popup.render();
 
             expect(view).toEqual(<AppearancePopup close={close} getSections={popup.getSections} noticeType="Notice of Appeal" />);
         });
         
         test('can build AppealRecordPopup', ()=>{            
-            let popup = mount(<InfoPopup show="true" type="appealrecord" close={close}/>).instance();
+            let popup = mount(<InfoPopupFactory show="true" type="appealrecord" close={close}/>).instance();
             let view = popup.render();
 
             expect(view).toEqual(<AppealRecordPopup close={close} getSections={popup.getSections}/>);
         });
         
         test('can build AppellantInitialPopup', ()=>{            
-            let popup = mount(<InfoPopup show="true" type="appellantinitial" close={close}/>).instance();
+            let popup = mount(<InfoPopupFactory show="true" type="appellantinitial" close={close}/>).instance();
             let view = popup.render();
 
             expect(view).toEqual(<AppellantInitialPopup close={close} getSections={popup.getSections}/>);
         });
         
         test('can build NoticeOfHearingPopup', ()=>{            
-            let popup = mount(<InfoPopup show="true" type="noticeofhearing" close={close}/>).instance();
+            let popup = mount(<InfoPopupFactory show="true" type="noticeofhearing" close={close}/>).instance();
             let view = popup.render();
 
             expect(view).toEqual(<NoticeOfHearingPopup close={close} getSections={popup.getSections}/>);
         });
         
         test('can build AppellantCourtOrderPopup', ()=>{            
-            let popup = mount(<InfoPopup show="true" type="appellantcourtorder" close={close}/>).instance();
+            let popup = mount(<InfoPopupFactory show="true" type="appellantcourtorder" close={close}/>).instance();
             let view = popup.render();
 
             expect(view).toEqual(<AppellantCourtOrderPopup close={close} getSections={popup.getSections}/>);
         });
 
         test('can build BookAnAppealPopup', ()=>{
-            let popup = mount(<InfoPopup show="true" type="bookappeal" close={close}/>).instance();
+            let popup = mount(<InfoPopupFactory show="true" type="bookappeal" close={close}/>).instance();
             let view = popup.render();
 
             expect(view).toEqual(<BookAnAppealPopup close={close} getSections={popup.getSections}/>);
         });
 
         test('can build ReviewPopup', ()=>{
-            let popup = mount(<InfoPopup show="true" type="review" close={close}/>).instance();
+            let popup = mount(<InfoPopupFactory show="true" type="review" close={close}/>).instance();
             let view = popup.render();
 
             expect(view).toEqual(<ReviewPopup close={close} getSections={popup.getSections}/>);
         });
 
         test('can build ReviewHearingPopup', ()=>{
-            let popup = mount(<InfoPopup show="true" type="reviewhearing" close={close}/>).instance();
+            let popup = mount(<InfoPopupFactory show="true" type="reviewhearing" close={close}/>).instance();
             let view = popup.render();
 
             expect(view).toEqual(<ReviewHearingPopup close={close} getSections={popup.getSections}/>);
         });
 
         test('can build ApplyForLeavePopup', ()=>{
-            let popup = mount(<InfoPopup show="true" type="applyforleave" close={close}/>).instance();
+            let popup = mount(<InfoPopupFactory show="true" type="applyforleave" close={close}/>).instance();
             let view = popup.render();
 
             expect(view).toEqual(<ApplyForLeavePopup close={close} getSections={popup.getSections}/>);
         });
 
         test('can build LeaveHearingDocumentsPopup', ()=>{
-            let popup = mount(<InfoPopup show="true" type="leavehearingdocs" close={close}/>).instance();
+            let popup = mount(<InfoPopupFactory show="true" type="leavehearingdocs" close={close}/>).instance();
             let view = popup.render();
 
             expect(view).toEqual(<LeaveHearingDocumentsPopup close={close} getSections={popup.getSections}/>);
         });
 
         test('can build LeaveHearingPopup', ()=>{
-            let popup = mount(<InfoPopup show="true" type="leavehearing" close={close}/>).instance();
+            let popup = mount(<InfoPopupFactory show="true" type="leavehearing" close={close}/>).instance();
             let view = popup.render();
 
             expect(view).toEqual(<LeaveHearingPopup close={close} getSections={popup.getSections}/>);
         });
 
         test('can build respondent DecisionPopup', ()=>{
-            let popup = mount( <InfoPopup
+            let popup = mount( <InfoPopupFactory
                 show="true"
                 type="respondentdecision"
                 close={close}
@@ -214,7 +214,7 @@ describe('InfoPopup', ()=> {
 
         test('can build appellant DecisionPopup', ()=>{
             let popup = mount(
-                <InfoPopup
+                <InfoPopupFactory
                     show="true"
                     type="decision"
                     close={close}
@@ -235,28 +235,28 @@ describe('InfoPopup', ()=> {
         });
 
         test('can build appellant CompleteAppealPopup', ()=>{
-            let popup = mount(<InfoPopup show="true" type="appellantcomplete" close={close} />).instance();
+            let popup = mount(<InfoPopupFactory show="true" type="appellantcomplete" close={close} />).instance();
             let view = popup.render();
 
             expect(view).toEqual(<CompleteAppealPopup close={close} getSections={popup.getSections} type="appellant" />);
         });
 
         test('can build respondent CompleteAppealPopup', ()=>{
-            let popup = mount(<InfoPopup show="true" type="respondentcomplete" close={close} />).instance();
+            let popup = mount(<InfoPopupFactory show="true" type="respondentcomplete" close={close} />).instance();
             let view = popup.render();
 
             expect(view).toEqual(<CompleteAppealPopup close={close} getSections={popup.getSections} type="respondent" />);
         });
 
         test('can build respondent RespondToAppealHearingPopup', ()=>{
-            let popup = mount(<InfoPopup show="true" type="respondentappealhearing" close={close} />).instance();
+            let popup = mount(<InfoPopupFactory show="true" type="respondentappealhearing" close={close} />).instance();
             let view = popup.render();
 
             expect(view).toEqual(<RespondToAppealHearingPopup close={close} getSections={popup.getSections} />);
         });
 
         test('can build respondent RespondToLeaveToAppealHearingPopup', ()=>{
-            let popup = mount(<InfoPopup show="true" type="respondentleavetoappealhearing" close={close} />).instance();
+            let popup = mount(<InfoPopupFactory show="true" type="respondentleavetoappealhearing" close={close} />).instance();
             let view = popup.render();
 
             expect(view).toEqual(<RespondToLeaveToAppealHearingPopup close={close} getSections={popup.getSections} />);
